Gameplay improvement
Added a possibility for playerTwo to make his move.
Added a console message with a name of the winner.
Added an increment for player score after win.
Added a loop for max. rounds = 5 -> (5x2 players)
Added an if statement that prints message on the console "no winner" if there won't be any winner after 5 turns.
Added a protection in Board.java which prevents a user for overriding a sign that already has been set (you cannot override your/your opponent's move).
Added a Tester.java class in the new package "tests" but it has to be written.
Added a temporary package "ttt" for the main classes.

The return of the method has been set to boolean (it needs a further analysis and maybe a rollback)